fix #47
