+++ backend/tests/test_persistance.py
@@ -1,8 +1,22 @@
-"""Tests for persistance.py"""
+"""Test get_watch_url"""
 from datetime import datetime
 import pytest
 from pingurl import persistance
+from pingurl.persistance import WatchedUrlNotFoundError
 from pingurl.models import WatchedUrl
+
+
+def test_get_watched_url_value_error():
+    """Tests if non int raises ValueError"""
+    with pytest.raises(ValueError):
+        persistance.get_watched_url("a")
+
+
+def test_get_watched_url_Watched_Url_Not_Found_Error():
+    """Tests if invalid url raises WatchedUrlNotFoundError"""
+    with pytest.raises(WatchedUrlNotFoundError):
+        persistance.get_watched_url(-1)
+
 
 def test_delete_watched_url_value_error():
     """test for delete wrong input type"""
